Fixed a bug That does not prevent files deleting when the upload fails Fixed typo

+++ uploader.sh
@@ -2,7 +2,7 @@
 #EDIT TOKEN:
 TOKEN="YOUR_TOKEN"
 FILE_NAME="$1"
-#ALSO MAKE SURE THAT CURL, RAR and JQ IS INSTALLED ON YOUR SYSTEM
+#ALSO MAKE SURE THAT CURL, RAR and JQ ARE INSTALLED ON YOUR SYSTEM
 upload() {
 	#$1 is filename
 	res=$(curl -X POST "https://bot.sapp.ir/$TOKEN/uploadFile" -H 'content-type: multipart/form-data' -F file=@"$1")
@@ -11,7 +11,7 @@ upload() {
 	if [[ "$ok" != "OK" ]]; then
 		echo "Error on file $1:"
 		jq .description <<<"$res"
-		delete=false
+		touch /tmp/SoroushUploader/.nodelete
 		return
 	fi
 	local id
@@ -27,7 +27,6 @@ rm -rf /tmp/SoroushUploader
 mkdir /tmp/SoroushUploader
 rar a -M0 -v100M /tmp/SoroushUploader/u.rar "$FILE_NAME" # You can also change the chunk size. Max upload size is 100MB
 #Then get the file names and upload eachone to server
-delete=true
 for filename in /tmp/SoroushUploader/*.rar; do
 	while [ "$(jobs | wc -l)" -ge 10 ]; do # Change 10 if needed
 		sleep 5
@@ -39,6 +38,4 @@ while [[ "$(jobs)" =~ "Running" ]]; do
 done
 echo "Done uploading files"
 #Delete the files after
-if [ "$delete" = true ]; then
-	rm -rf /tmp/SoroushUploader
-fi
+[ -f /tmp/SoroushUploader/.nodelete ] && rm -rf /tmp/SoroushUploader
